requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
h5py==3.8.0
kiwisolver==1.4.8
matplotlib==3.7.1
numpy==1.23.5
opencv-python-headless==4.8.0.74
packaging==25.0
Pillow==9.5.0
pyparsing==3.2.3
PyQt5==5.15.7
PyQt5-Qt5==5.15.17
PyQt5-sip==12.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0